agents/decision/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
redis==5.2.0
oracledb==2.4.1
google-generativeai==0.8.3
httpx==0.27.2
prometheus-client==0.21.0
